+++ cmake_setup/cpp_cuda/simpleAddGPU.h
@@ -1,7 +1,7 @@
 #ifndef _SIMPLEADD_GPU
 #define _SIMPLEADD_GPU
 
-__global__ void randomize_cuda(float* a,float* b, float* c,long N);
-__global__ void add_cuda(float* a,float* b,float* c,long N);
+__global__ void randomize_cuda(double* a,double* b, double* c,long N);
+__global__ void add_cuda(double* a,double* b,double* c,long N);
 
 #endif
